Remove duplicate implementations in the analysis app #1799
- SignificanceFilterNodeMixin holds the six members ClassificationsNode and
  ClinVarNode had a copy of each
- _get_cached_label_count moves up to AbstractCohortBasedNode, reached through
  the _get_cohort() each subclass already implements. TrioNode keeps its
  has_input() guard as an override
- _source_grid_export shares what cohort_grid_export and sample_grid_export had
  in common, and node_grid_export joins them on EXPORT_TYPES validation
- _family_inheritance holds what the trio and quad inheritance hierarchies share:
  the zygosity constants and display helper, and the whole of comp het, which
  differed only in reaching the family through .trio vs .quad
