+++ analysis/models/nodes/filters/clinvar_node.py
@@ -9,6 +9,7 @@ from django.db.models import Q
 from analysis.models.enums import NodeMatchInput
 from analysis.models.nodes.analysis_node import AnalysisNode
 from analysis.models.nodes.node_display import NodeChip, NodeIcon, significance_chips
+from analysis.models.nodes.significance_filter_mixin import SignificanceFilterNodeMixin
 from annotation.models.models_enums import ClinVarOncogenicity, ClinVarPathogenicity, ClinVarReviewStatus
 from classification.enums import SomaticClinicalSignificance
 from snpdb.models.models_enums import AlleleOriginFilterDefault
@@ -73,7 +74,7 @@ SOMATIC_TIER_LABELS = SomaticClinicalSignificance.LABELS | {NO_SOMATIC_TIER: "No
 SOMATIC_TIER_SHORT_LABELS = SomaticClinicalSignificance.SHORT_LABELS | {NO_SOMATIC_TIER: "None"}
 
 
-class ClinVarNode(AnalysisNode):
+class ClinVarNode(SignificanceFilterNodeMixin, AnalysisNode):
     """ What ClinVar says about a variant - every pill on means "has a ClinVar record" """
     node_input = models.CharField(max_length=1, choices=NodeMatchInput.choices,
                                   default=NodeMatchInput.PARENT_MATCHING)
@@ -130,31 +131,6 @@ class ClinVarNode(AnalysisNode):
         'oncogenicity_benign': ClinVarOncogenicity.BENIGN,
         'oncogenicity_none': NO_CLINVAR_CALL,
     }
-
-    @property
-    def min_inputs(self):
-        return self.max_inputs
-
-    @property
-    def max_inputs(self):
-        if self.node_input == NodeMatchInput.MATCHING_VARIANTS:
-            return 0
-        return 1
-
-    @property
-    def allele_origin_filter(self) -> AlleleOriginFilterDefault:
-        return AlleleOriginFilterDefault(self.allele_origin)
-
-    @property
-    def germline_enabled(self) -> bool:
-        return self.allele_origin_filter != AlleleOriginFilterDefault.SOMATIC
-
-    @property
-    def somatic_enabled(self) -> bool:
-        return self.allele_origin_filter != AlleleOriginFilterDefault.GERMLINE
-
-    def _selected_values(self, field_values: dict) -> list:
-        return [value for field, value in field_values.items() if getattr(self, field)]
 
     def _filtering_values(self, field_values: dict) -> list:
         """ A row with every pill on is 'everything ClinVar has', so it applies no filter """
